Moved parsing logic to separate package
